rename query_bar folder to query_field
