+++ x-pack/plugins/security_solution/public/detection_engine/rule_creation_ui/components/query_field/default_queries.ts
