+++ x-pack/plugins/security_solution/public/detection_engine/rule_creation_ui/components/query_field/query_field.tsx
@@ -65,7 +65,7 @@ const savedQueryToFieldValue = (savedQuery: SavedQuery): FieldValueQueryBar => (
   title: savedQuery.attributes.title,
 });
 
-export const QueryBarDefineRule = ({
+export const QueryField = ({
   defaultSavedQuery,
   dataTestSubj,
   field,
