+++ x-pack/plugins/security_solution/public/detection_engine/rule_creation_ui/components/step_define_rule/use_persistent_query.ts
@@ -22,7 +22,7 @@ import {
   DEFAULT_KQL_QUERY_FIELD_VALUE,
   DEFAULT_THREAT_MATCH_KQL_QUERY_FIELD_VALUE,
   type FieldValueQueryBar,
-} from '../query_bar';
+} from '../query_field';
 
 const EQL_QUERY_LANGUAGE = 'eql';
 const ESQL_QUERY_LANGUAGE = 'esql';
