+++ x-pack/plugins/security_solution/public/detection_engine/rule_management/components/rule_details/three_way_diff/final_edit/fields/kql_query/kql_query_edit.tsx
@@ -12,7 +12,7 @@ import { EuiButtonEmpty } from '@elastic/eui';
 import type { DataViewBase } from '@kbn/es-query';
 import { schema } from '../../../../../../../rule_creation_ui/components/step_define_rule/schema';
 import { HiddenField, UseField } from '../../../../../../../../shared_imports';
-import { QueryBarDefineRule } from '../../../../../../../rule_creation_ui/components/query_bar';
+import { QueryField } from '../../../../../../../rule_creation_ui/components/query_field';
 import * as stepDefineRuleI18n from '../../../../../../../rule_creation_ui/components/step_define_rule/translations';
 import type { DiffableRule } from '../../../../../../../../../common/api/detection_engine';
 import type { SetRuleQuery } from '../../../../../../../../detections/containers/detection_engine/rules/use_rule_from_timeline';
@@ -58,7 +58,7 @@ export function KqlQueryEdit({
             <ImportTimelineQueryButton handleOpenTimelineSearch={toggleIsTimelineSearchOpen} />
           ),
         }}
-        component={QueryBarDefineRule}
+        component={QueryField}
         componentProps={{
           indexPattern: dataView ?? DEFAULT_DATA_VIEW,
           isLoading,
