+++ x-pack/plugins/security_solution/public/detection_engine/rule_management/components/rule_details/three_way_diff/final_edit/fields/kql_query/kql_query_edit_form.tsx
@@ -10,7 +10,7 @@ import type { Type } from '@kbn/securitysolution-io-ts-alerting-types';
 import type { FormData, FormSchema } from '../../../../../../../../shared_imports';
 import { schema } from '../../../../../../../rule_creation_ui/components/step_define_rule/schema';
 import { RuleFieldEditFormWrapper } from '../rule_field_edit_form_wrapper';
-import type { FieldValueQueryBar } from '../../../../../../../rule_creation_ui/components/query_bar';
+import type { FieldValueQueryBar } from '../../../../../../../rule_creation_ui/components/query_field';
 import {
   KqlQueryLanguage,
   KqlQueryType,
