+++ x-pack/plugins/security_solution/public/detection_engine/rule_management_ui/components/rules_table/__mocks__/mock.ts
@@ -16,7 +16,7 @@ import {
   DataSourceType,
   AlertSuppressionDurationType,
 } from '../../../../../detections/pages/detection_engine/rules/types';
-import type { FieldValueQueryBar } from '../../../../rule_creation_ui/components/query_bar';
+import type { FieldValueQueryBar } from '../../../../rule_creation_ui/components/query_field';
 import { fillEmptySeverityMappings } from '../../../../../detections/pages/detection_engine/rules/helpers';
 import { getThreatMock } from '../../../../../../common/detection_engine/schemas/types/threat.mock';
 import {
