+++ x-pack/plugins/security_solution/public/timelines/components/timeline/query_bar/eql/index.tsx
@@ -17,7 +17,6 @@ import { useDeepEqualSelector } from '../../../../../common/hooks/use_selector';
 import { SourcererScopeName } from '../../../../../sourcerer/store/model';
 import { EqlQueryEdit } from '../../../../../detection_engine/rule_creation/components/eql_query_edit';
 import type { FieldValueQueryBar } from '../../../../../detection_engine/rule_creation_ui/components/query_bar';
-
 import type { FormSchema, FormSubmitHandler } from '../../../../../shared_imports';
 import { Form, UseField, useForm } from '../../../../../shared_imports';
 import { timelineActions } from '../../../../store';
